requirements.txt
google-generativeai>=0.3.0
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0